pyproject.toml
[tool.poetry]
name = "sy-low-value"
version = "0.1.0"
description = "顺义低值可回收物平台"
authors = ["zhuangbaojun <zhuangbaojun@ztbory.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.3"
django-ninja = "^0.18.0"
django-extensions = "^3.1.5"
PyJWT = "^2.3.0"
django-cors-headers = "^3.13.0"
sentry-sdk = "^1.5.7"
gevent = "^21.12.0"
python-dateutil = "^2.8.2"
requests = "^2.27.1"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.3"
shortuuid = "^1.0.8"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"

[tool.isort]
# multi_line_output = 3
# include_trailing_comma = true
# force_grid_wrap = 0
line_length = 120

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
